+++ CHANGELOG.md
@@ -7,9 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [2.0.0] - 2022-03-15
-### Added
-- `NTBLMagic.project_git_dir` config for the git directory
-
 ### Changed
 - Upgrade `ipython` to `^7.31.1` for security fix
 - Upgrade `numpy` to `^1.22.2` for security fix
@@ -18,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `%ntbl` commands that interact with `git` directly. This will need to be added to the planar-ally api now. 
   - Remove `%ntbl diff project`
   - Remove `%ntbl status project`
+- Remove no longer used `NTBLMagic` config values: `git_user_name` and `git_user_email`
 
 ## [1.2.4] - 2021-09-01
 ### Added